+++ README.md
@@ -21,7 +21,7 @@
 - [X] 로또 1장의 가격은 1,000원이다.
 - [X] 당첨 번호와 보너스 번호를 입력받는다.
 - [X] 사용자가 구매한 로또 번호와 당첨 번호를 비교하여 당첨 내역 및 수익률을 출력하고 로또 게임을 종료한다.
-- [ ] 사용자가 잘못된 값을 입력할 경우 `IllegalArgumentException`을 발생시키고, "[ERROR]"로 시작하는 에러 메시지를 출력 후 그 부분부터 입력을 다시 받는다.
+- [X] 사용자가 잘못된 값을 입력할 경우 `IllegalArgumentException`을 발생시키고, "[ERROR]"로 시작하는 에러 메시지를 출력 후 그 부분부터 입력을 다시 받는다.
     - `Exception`이 아닌 `IllegalArgumentException`, `IllegalStateException` 등과 같은 명확한 유형을 처리한다.
 
 ### **입출력 요구 사항**
@@ -239,7 +239,7 @@
 
 ### 예외 발생 시
 - [X] [ERROR] 접두사와 함께 에러 메시지 출력
-- [ ] 해당 입력부터 다시 받기
+- [X] 해당 입력부터 다시 받기
 ---
 
 ## 🤔 생각하면서 진행해본 내용 (추후 생각 나는대로 작성 예정)
