+++ src/test/java/lotto/service/LottoPurchaseServiceTest.java
@@ -2,7 +2,7 @@ package lotto.service;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import lotto.service.dto.LottoPurchaseResult;
+import lotto.dto.LottoPurchaseResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
